Moved CpiFileEntity from net.corda.libs.cpi.datamodel.entities to net.corda.libs.cpi.datamodel.entities.internal so the class is not exposed by OSGI.
