+++ libs/virtual-node/cpi-datamodel/src/main/kotlin/net/corda/libs/cpi/datamodel/entities/internal/CpkFileEntity.kt
@@ -1,4 +1,4 @@
-package net.corda.libs.cpi.datamodel.entities
+package net.corda.libs.cpi.datamodel.entities.internal
 
 import java.time.Instant
 import javax.persistence.Column
